# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SewoApp API Documentation",
        "description": "Dokumentasi API untuk SewoApp",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Mendapatkan daftar semua pesanan",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "Daftar pesanan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Membuat pesanan baru",
                "operationId": "2123ba58dadb22bd6f2c8081af562230",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_customer",
                                    "order_date",
                                    "start_date",
                                    "finish_date",
                                    "status",
                                    "total_price",
                                    "payment_method"
                                ],
                                "properties": {
                                    "id_customer": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "order_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-14"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-15"
                                    },
                                    "finish_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-20"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500000
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "transfer"
                                    },
                                    "payment_gateway_ref": {
                                        "type": "string",
                                        "example": "TRX123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pesanan berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Mendapatkan detail pesanan",
                "operationId": "88d27387b1e409e075c18609d4a0c300",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pesanan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail pesanan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pesanan tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Mengupdate pesanan",
                "operationId": "38a1b85e8294795a1b0dc438a36e319f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pesanan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_customer": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "order_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-14"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-15"
                                    },
                                    "finish_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-20"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "completed"
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500000
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "transfer"
                                    },
                                    "payment_gateway_ref": {
                                        "type": "string",
                                        "example": "TRX123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pesanan berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pesanan tidak ditemukan"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Menghapus pesanan",
                "operationId": "bab875a1ed962fb33565ab9b48c0d494",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pesanan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Pesanan berhasil dihapus"
                    },
                    "404": {
                        "description": "Pesanan tidak ditemukan"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Order": {
                "required": [
                    "id_customer",
                    "order_date",
                    "start_date",
                    "finish_date",
                    "status",
                    "total_price",
                    "payment_method"
                ],
                "properties": {
                    "id_order": {
                        "description": "ID pesanan",
                        "type": "integer",
                        "example": 1
                    },
                    "id_customer": {
                        "description": "ID customer",
                        "type": "integer",
                        "example": 1
                    },
                    "order_date": {
                        "description": "Tanggal pemesanan",
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-14"
                    },
                    "start_date": {
                        "description": "Tanggal mulai sewa",
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-15"
                    },
                    "finish_date": {
                        "description": "Tanggal selesai sewa",
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-20"
                    },
                    "status": {
                        "description": "Status pesanan",
                        "type": "string",
                        "example": "pending"
                    },
                    "total_price": {
                        "description": "Total harga",
                        "type": "number",
                        "format": "float",
                        "example": 500000
                    },
                    "payment_method": {
                        "description": "Metode pembayaran",
                        "type": "string",
                        "example": "transfer"
                    },
                    "payment_gateway_ref": {
                        "description": "Referensi payment gateway",
                        "type": "string",
                        "example": "TRX123456"
                    },
                    "created_at": {
                        "description": "Tanggal dibuat",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-14T10:00:00Z"
                    },
                    "updated_at": {
                        "description": "Tanggal diupdate",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-14T10:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}